Move app info view sub components to separate folders
