+++ desktop/packages/mullvad-vpn/src/renderer/components/views/app-info/components/index.ts
@@ -1,3 +1,3 @@
-export * from './VersionListItem';
-export * from './ChangelogListItem';
-export * from './UpdateAvailableListItem';
+export * from './version-list-item';
+export * from './changelog-list-item';
+export * from './update-available-list-item';
