+++ desktop/packages/mullvad-vpn/src/renderer/components/views/app-info/components/version-list-item/index.ts
@@ -1,0 +1,1 @@
+export * from './VersionListItem';
